Add ability to query by zip code to forecast service
Note that I chose to stub out the database calls for the zip code
test cases. This is due in large part to the extra setup hassle on the
CI side since pre-loading the zip codes requires executing
a (long-running) rake task before each run of the build.

Consider it a compromise in the face of a looming deadline.